token_ai_tracker/requirements.txt
web3
pandas
sqlalchemy
requests
python-dotenv
